fixed importer script for magazine pages
